docs: add KDocs of new MVVM pattern implementation
