+++ application/src/main/kotlin/utils/PageType.kt
@@ -1,5 +1,8 @@
 package utils
 
+/**
+ * Enum of pages on main screen of application.
+ */
 enum class PageType {
 	HOME_PAGE,
 	GRAPH_VIEW_PAGE,
