+++ application/src/main/kotlin/utils/VertexIDType.kt
@@ -1,5 +1,13 @@
 package utils
 
+/**
+ * Vertex identifier type on application representation.
+ *
+ * __Important:__ It's needed to valid execution of application and using on graphs objects.
+ *
+ * @property title simple string of name element
+ * @constructor Create Vertex ID type with custom label
+ */
 enum class VertexIDType(val title: String) {
 	INT_TYPE(Int::class.simpleName.toString()),
 	STRING_TYPE(String::class.simpleName.toString());
