+++ application/src/main/kotlin/viewmodels/dialogs/CreateNewGraphDialogViewModel.kt
@@ -1,7 +1,7 @@
 package viewmodels.dialogs
 
 import androidx.compose.runtime.mutableStateOf
-import utils.SaveType
+import utils.GraphSavingType
 import utils.VertexIDType
 import viewmodels.pages.HomePageViewModel
 
@@ -10,5 +10,5 @@ class CreateNewGraphDialogViewModel(val homePageViewModel: HomePageViewModel) {
 	val selectedVertexTypeID = mutableStateOf(VertexIDType.INT_TYPE)
 	val isGraphWeighted = mutableStateOf(false)
 	val isGraphDirected = mutableStateOf(false)
-	val selectedSaveType = mutableStateOf(SaveType.LOCAL_FILE)
+	val selectedSaveType = mutableStateOf(GraphSavingType.LOCAL_FILE)
 }
